Add files via upload
This is a script that generates a correlation matrix as a heatmap. Within each tile of the heatmap are symbols for fdr adjusted p-values, correlation coefficient, which correspond to the color of the tile, and the sample size of the correlation. You may not be interested in using every element of this plot for your own purposes but it demonstrates the specificity with which such graphics can be generated using the ggplot2 library in R. This graphic even includes a subplot that is layered within the heatmap as a way to add an additional legend to the plot to orient a reader. One note is that the text and labels within the plot do not automatically scale to the size of the graph so depending on the size of your desired graph, tweaking of the text size and position may be necessary. So don't worry if it looks like a jumbled mess at first. Once you click zoom and adjust the size you can see all the elements within the dimensions that work best for it.